requirements.txt
python-gitlab~=4.5.0
tqdm~=4.66.4